+++ apps/web/src/app/[locale]/[...catchAll404]/page.tsx
@@ -1,0 +1,6 @@
+import { notFound } from 'next/navigation';
+import NotFound from '../not-found';
+
+export default function CatchAllPage() {
+  return <NotFound />;
+}
