+++ apps/web/src/app/[locale]/layout.tsx
@@ -8,6 +8,46 @@ import { Navigation } from '~/components/navigation';
 import { getStaticParams } from '~/locales/server';
 import { Providers } from './providers';
 
+import { OG_URL, tagline } from '../metadata';
+
+export const metadata = {
+  metadataBase: new URL(OG_URL),
+  title: 'TypeHero',
+  robots: {
+    index: true,
+    follow: true,
+  },
+  description: tagline,
+  openGraph: {
+    title: 'TypeHero',
+    description: tagline,
+    siteName: 'TypeHero',
+    images: [
+      {
+        url: `${OG_URL}/api/default`,
+        width: 1920,
+        height: 1080,
+      },
+    ],
+    locale: 'en-US',
+    type: 'website',
+  },
+  twitter: {
+    title: 'TypeHero',
+    card: 'summary_large_image',
+    images: [
+      {
+        url: `${OG_URL}/api/default`,
+        width: 1920,
+        height: 1080,
+      },
+    ],
+  },
+  icons: {
+    shortcut: '/favicon.ico',
+  },
+};
+
 const inter = Inter({ subsets: ['latin'] });
 
 export function generateStaticParams() {
